데이터 로드 구조화,

+++ console_project/Time.cpp
@@ -2,11 +2,11 @@
 #include <Windows.h>
 #include "ConsoleDatas.h"
 
-unsigned long long prev_time = 0;
+int prev_time = 0;
 
 void TimeInit()
 {
-	prev_time = static_cast<unsigned long long>(timeGetTime()) - target_delta_time;
+	prev_time = static_cast<int>(timeGetTime()) - target_delta_time;
 }
 
 void AsyncFrame()
@@ -16,9 +16,9 @@ void AsyncFrame()
 	// 3. 목표 프레임에 비해 너무 많으면 sleep으로 대기
 	// 4. 목표 프레임 만큼 현재 시간 증가
 	// 게임 내 독자적인 시간을 가진다.
-	// 내 세상의 시간은 이래.
+	// 내 세상의 시간은 이렇게 흘러. 존나 멋있음.
 
-	int now_time = static_cast<unsigned long long>(timeGetTime());
+	int now_time = static_cast<int>(timeGetTime());
 	int used_time = now_time - prev_time;
 
 	if (used_time < target_delta_time)
